requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
yarg==0.1.9
django-filter==23.2
django-admin-searchable-dropdown==1.2
Pillow==10.0.0
djangorestframework==3.14.0
pytz==2023.3
serializers ==0.2.4
cffi==1.15.1
cryptography==41.0.3
defusedxml==0.7.1
django-templated-mail==1.1.1
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
oauthlib==3.2.2
pycparser==2.21
pyjwt==2.8.0
python3-openid==3.2.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
django-chartjs==2.3.0
faker==19.6.1
python-dateutil==2.8.2
gunicorn==21.2.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
django-redis==5.3.0
redis==5.0.0
celery~=5.3.4
django-crispy-forms==2.0
crispy-bootstrap5==0.7